processor/requirements.txt
pytest
numpy
scipy
pyserial
